+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bump pypa/gh-action-pypi-publish from 1.9.0 to 1.10.2 (by [@dependabot](https://github.com/apps/dependabot) in [#22](https://github.com/stumpylog/tika-client/pull/22))
 - Update `pre-commit` to 4.0.1 ([#23](https://github.com/stumpylog/tika-client/pull/23))
 - Use pytest fixtures effectively ([#24](https://github.com/stumpylog/tika-client/pull/24))
+- Use pytest-docker in place of manual Docker ([#26](https://github.com/stumpylog/tika-client/pull/26))
 
 ## [0.6.0] - 2024-07-18